# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-Commerce Demo API",
    "version": "1.0.0",
    "description": "花卉電商網站 REST API"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Id"
      }
    }
  },
  "paths": {
    "/api/admin/orders": {
      "get": {
        "summary": "後台訂單列表",
        "tags": [
          "Admin Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "paid",
                "failed"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "orders": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "order_no": {
                                "type": "string"
                              },
                              "user_id": {
                                "type": "string"
                              },
                              "recipient_name": {
                                "type": "string"
                              },
                              "recipient_email": {
                                "type": "string"
                              },
                              "total_amount": {
                                "type": "integer"
                              },
                              "status": {
                                "type": "string"
                              },
                              "created_at": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "integer"
                            },
                            "page": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/orders/{id}": {
      "get": {
        "summary": "後台訂單詳情",
        "tags": [
          "Admin Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "order_no": {
                          "type": "string"
                        },
                        "user_id": {
                          "type": "string"
                        },
                        "recipient_name": {
                          "type": "string"
                        },
                        "recipient_email": {
                          "type": "string"
                        },
                        "recipient_address": {
                          "type": "string"
                        },
                        "total_amount": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "product_id": {
                                "type": "string"
                              },
                              "product_name": {
                                "type": "string"
                              },
                              "product_price": {
                                "type": "integer"
                              },
                              "quantity": {
                                "type": "integer"
                              }
                            }
                          }
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "訂單不存在"
          }
        }
      }
    },
...
    "/api/orders": {
      "post": {
        "summary": "從購物車建立訂單",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "recipientName",
                  "recipientEmail",
                  "recipientAddress"
                ],
                "properties": {
                  "recipientName": {
                    "type": "string"
                  },
                  "recipientEmail": {
                    "type": "string",
                    "format": "email"
                  },
                  "recipientAddress": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "訂單建立成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "order_no": {
                          "type": "string"
                        },
                        "total_amount": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "product_name": {
                                "type": "string"
                              },
                              "product_price": {
                                "type": "integer"
                              },
                              "quantity": {
                                "type": "integer"
                              }
                            }
                          }
                        },
                        "created_at": {
                          "type": "string"
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "購物車為空或庫存不足或收件資訊缺失"
          }
        }
      },
      "get": {
        "summary": "自己的訂單列表",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "orders": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "order_no": {
                                "type": "string"
                              },
                              "total_amount": {
                                "type": "integer"
                              },
                              "status": {
                                "type": "string"
                              },
                              "created_at": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders/{id}/ecpay/checkout": {
      "post": {
        "summary": "產生綠界 AIO 付款表單參數",
        "description": "回傳前端需自動送出（POST）到綠界付款頁的 action_url 與表單欄位，僅支援信用卡一次付清；只有 pending 訂單可呼叫。",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "400": {
            "description": "訂單狀態不是 pending"
          },
          "404": {
            "description": "訂單不存在"
          },
          "500": {
            "description": "綠界環境變數未設定"
          }
        }
      }
    },
    "/api/orders/{id}/ecpay/confirm": {
      "post": {
        "summary": "主動查詢綠界交易並確認付款結果",
        "description": "呼叫綠界 QueryTradeInfo/V5，驗證 CheckMacValue 與金額後才更新訂單狀態；本機環境不依賴 ReturnURL 回呼，此端點是付款結果確認的唯一依據。對已是最終狀態（paid/failed）的訂單直接回傳現況，不重複查詢。",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查詢完成（可能是 paid／failed，或仍是 pending 待稍後再查）"
          },
          "400": {
            "description": "尚未建立綠界付款交易"
          },
          "404": {
            "description": "訂單不存在"
          },
          "500": {
            "description": "綠界環境變數未設定"
          },
          "502": {
            "description": "呼叫綠界失敗、CheckMacValue 驗證失敗，或回應與訂單不符"
          }
        }
      }
    },
...
  },
...
}